Stop using dash as delimiter for hash computation
While this techincally is a breaking change, the name template feature is
hopefully new enough that we can get away with it.
Currently, there is a flaw in the hash computation that allows conflicts to occur.
This adds a test for this edge case and fixes the bug by using a delimeter character
that is not allowed in kubernetes object names.
The semicolon character was chosen since this is what ingress-gce uses to avoid this issue
[here](https://github.com/kubernetes/ingress-gce/blob/c4fd082b87afdc5fa294920477eca368db269d87/pkg/utils/namer/namer.go#L492)