Support getting zones using svcneg resources.
This change adds support for getting NEG zones
using svcneg resources created by GKE. This is
controlled via a flag `use-svcneg`.

See https://cloud.google.com/kubernetes-engine/docs/how-to/standalone-neg#validate-neg-creation.
This is only supported for GKE versions newer than
1.18.6-gke.6400.

This approach is recommended by the GKE networking
team as it is updated faster than the status
annotation on the service.

This change was tested for a few days on
test GKE clusters used by the Vertex AI team.